# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StoreKit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: collectionSlug
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: newArrival
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, featured]
        - name: minPrice
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: integer
        - name: sizes
          in: query
          schema:
            type: string
        - name: colors
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /collections/{slug}:
    get:
      operationId: getCollectionBySlug
      tags: [collections]
      summary: Get collection by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin-products]
      summary: Admin - list all products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: collectionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductListResponse"
    post:
      operationId: adminCreateProduct
      tags: [admin-products]
      summary: Admin - create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    get:
      operationId: adminGetProduct
      tags: [admin-products]
      summary: Admin - get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: adminUpdateProduct
      tags: [admin-products]
      summary: Admin - update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin-products]
      summary: Admin - delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/collections:
    get:
      operationId: adminListCollections
      tags: [admin-collections]
      summary: Admin - list all collections
      responses:
        "200":
          description: Collections list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: adminCreateCollection
      tags: [admin-collections]
      summary: Admin - create collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /admin/collections/{id}:
    patch:
      operationId: adminUpdateCollection
      tags: [admin-collections]
      summary: Admin - update collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
    delete:
      operationId: adminDeleteCollection
      tags: [admin-collections]
      summary: Admin - delete collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    ProductImage:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        alt:
          type: string
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
      required: [id, url, alt, isPrimary, sortOrder]
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        size:
          type: string
        color:
          type: string
        colorHex:
          type: string
        sku:
          type: string
        stock:
          type: integer
        price:
          type: integer
        compareAtPrice:
          type: ["integer", "null"]
      required: [id, size, color, colorHex, sku, stock, price]
    Product:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        collectionIds:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, draft, archived]
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        basePrice:
          type: integer
        compareAtPrice:
          type: ["integer", "null"]
        material:
          type: ["string", "null"]
        careInstructions:
          type: ["string", "null"]
        weight:
          type: ["integer", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - name
        - description
        - shortDescription
        - images
        - variants
        - collectionIds
        - tags
        - status
        - isFeatured
        - isNewArrival
        - basePrice
        - createdAt
        - updatedAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, pageSize, totalPages]
    AdminProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, pageSize, totalPages]
    Collection:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        productCount:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - name
        - description
        - isFeatured
        - sortOrder
        - productCount
        - createdAt
        - updatedAt
...
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        basePrice:
          type: integer
        compareAtPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, draft, archived]
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        material:
          type: ["string", "null"]
        careInstructions:
          type: ["string", "null"]
        weight:
          type: ["integer", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        collectionIds:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              alt:
                type: string
              isPrimary:
                type: boolean
              sortOrder:
                type: integer
            required: [url, alt, isPrimary, sortOrder]
        variants:
          type: array
          items:
            type: object
            properties:
              size:
                type: string
              color:
                type: string
              colorHex:
                type: string
              sku:
                type: string
              stock:
                type: integer
              price:
                type: integer
              compareAtPrice:
                type: ["integer", "null"]
            required: [size, color, colorHex, sku, stock, price]
      required: [name, slug, description, shortDescription, basePrice, status]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        basePrice:
          type: integer
        compareAtPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, draft, archived]
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
        material:
          type: ["string", "null"]
        careInstructions:
          type: ["string", "null"]
        weight:
          type: ["integer", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        collectionIds:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              alt:
                type: string
              isPrimary:
                type: boolean
              sortOrder:
                type: integer
            required: [url, alt, isPrimary, sortOrder]
        variants:
          type: array
          items:
            type: object
            properties:
              size:
                type: string
              color:
                type: string
              colorHex:
                type: string
              sku:
                type: string
              stock:
                type: integer
              price:
                type: integer
              compareAtPrice:
                type: ["integer", "null"]
            required: [size, color, colorHex, sku, stock, price]
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
      required: [name, slug, description]
    UpdateCollectionBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
...