# lib/api-zod/src/generated/types/adminProductListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StoreKit API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface AdminProductListResponse {
  products: Product[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}